requirements.txt
torch>=1.12
torchvision>=0.13
scikit-learn>=1.0
pandas>=1.3
pyarrow>=8.0
matplotlib>=3.5
opencv-python>=4.5
torchcam>=0.3.2
Pillow>=9.0
tqdm>=4.64